requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
langchain
langchain-community
openai
tiktoken
unstructured
pdfminer==20191125 
pdfminer.six==20221105 
pillow_heif 
unstructured_inference 
youtube-transcript-api 
pytube 
sentence-transformers
langchain-huggingface
groq
langchain-chroma>=0.1.2
python-dotenv
chroma-hnswlib